requirements.txt
requests
openai
python-dotenv
typer
pytest
tenacity
flask
